Move response to clip.js
